v1 hygiene: CLI/bench/test warning + sanitizer fixes
Conversion / sign-conversion cleanup across the non-library sources now
that -Wconversion is on by default.  Most sites are size_t / int crossings
in printf paths and loop counters; added explicit casts rather than
changing variable types.

c3d_test.c — ASan/UBSan fixes:
  - make_geometric_freqs: UB shift-by-64 when n_symbols == 65 (the
    sign-predictive alphabet size); cap at 63 instead.  Geometric
    ratio saturates in the top handful of symbols, which is what the
    later normalisation already expects.
  - test_panic_hook annotated _Noreturn to match actual behaviour
    (longjmp or abort) and silence clang's -Wmissing-noreturn.
  - EXPECT_PANIC macro now brackets the longjmp region with
    __lsan_disable / __lsan_enable (compile-time gated on
    __has_feature(address_sanitizer) / __SANITIZE_ADDRESS__).  The
    panic-path allocations legitimately leak under setjmp unwind; LSan
    noise obscured the 3 real leaks I was chasing.
  - test_quant_roundtrip: float loop counter replaced with integer
    counter (0.123f increments accumulated drift after 16k iterations).
  - test_ctx_with_freq_tables: CHECK(ctx != NULL) promoted to
    c3d_assert to fix clang-analyzer's null-deref path through the
    subsequent CHECK(ctx->has_freq_tables).
  - Added braces on compact triple-for loops flagged by
    readability-misleading-indentation (scatter round-trip test,
    downsample spot-check) — correctness unchanged, reads clearer.

c3d_bench.c:
  - (char *)(uintptr_t)<string-literal> casts on the tthresh argv so the
    cast-qual/discarded-qualifiers warnings go away without disabling
    the checks globally.
  - Explicit (size_t) casts on the W*H chunk-stride arithmetic in the
    video codec paths; libde265's decoded_image stride/plane arithmetic
    needed the same treatment.
  - print_summary: precompute n as double once per code point, drop the
    64 implicit size_t→double conversions in the table body.

c3d_bench_par.c:
  - n_threads is int (atoi), casted to size_t at every arithmetic site
    that mixes with n_paths / paths_cap.
  - Summary-printer's MB/s / PSNR aggregates cast bytes/chunks to double
    at the arithmetic (not inside printf %f).

c3d_progressive.c, c3d_perf.c, c3d_train_dz.c — same shape of fixes;
sign/widening casts at the arithmetic, explicit (double) where size_t
feeds a printf format.

Also c3d_progressive.c: aligned_alloc(32, st.st_size) now rounds st_size
up to a multiple of 32 (ASan was rejecting the raw value on non-aligned
file sizes).

+++ c3d_bench_par.c
@@ -124,15 +124,17 @@ int main(int argc, char **argv) {
                n_obs, c3d_ctx_serialized_size(ctx));
     }
 
-    pthread_t *thr = calloc(n_threads, sizeof *thr);
-    worker_arg *args = calloc(n_threads, sizeof *args);
+    pthread_t *thr = calloc((size_t)n_threads, sizeof *thr);
+    worker_arg *args = calloc((size_t)n_threads, sizeof *args);
     /* Round-robin assignment. */
-    size_t per = (n_paths + n_threads - 1) / n_threads;
+    size_t per = (n_paths + (size_t)n_threads - 1) / (size_t)n_threads;
     for (int t = 0; t < n_threads; ++t) {
         args[t].tid = t;
-        args[t].paths = &paths[t * per];
-        args[t].n_paths = (t * per >= n_paths) ? 0
-                        : (((t + 1) * per <= n_paths) ? per : n_paths - t * per);
+        args[t].paths = &paths[(size_t)t * per];
+        args[t].n_paths = ((size_t)t * per >= n_paths) ? 0
+                        : ((size_t)(t + 1) * per <= n_paths
+                               ? per
+                               : n_paths - (size_t)t * per);
         args[t].target_ratio = target;
         args[t].ctx = ctx;
     }
@@ -147,23 +149,23 @@ int main(int argc, char **argv) {
            "tid", "chunks", "bytes_in", "bytes_enc", "MB/s_in", "PSNR_avg");
     for (int t = 0; t < n_threads; ++t) {
         if (args[t].n_paths == 0) continue;
-        double mbps = (args[t].total_bytes_in / (1024.0*1024.0)) / args[t].elapsed_s;
+        double mbps = ((double)args[t].total_bytes_in / (1024.0*1024.0)) / args[t].elapsed_s;
         printf("%-3d %-7zu %12zu %10zu %10.1f %.2f\n",
                t, args[t].n_paths, args[t].total_bytes_in, args[t].total_bytes_enc,
                mbps, args[t].avg_psnr);
         agg_bytes_in  += args[t].total_bytes_in;
         agg_bytes_enc += args[t].total_bytes_enc;
-        psnr_sum += args[t].avg_psnr * args[t].n_paths;
+        psnr_sum += args[t].avg_psnr * (double)args[t].n_paths;
         n_done += args[t].n_paths;
     }
-    double agg_mbps = (agg_bytes_in / (1024.0*1024.0)) / total;
+    double agg_mbps = ((double)agg_bytes_in / (1024.0*1024.0)) / total;
     printf("---\n");
     printf("aggregate: %zu chunks in %.2fs, %.1f MB/s in (%.2fx scale-up vs ideal=%dx)\n",
            n_done, total, agg_mbps,
-           agg_mbps / ((agg_bytes_in / (1024.0*1024.0)) / args[0].elapsed_s) /* approx scale */,
+           agg_mbps / (((double)agg_bytes_in / (1024.0*1024.0)) / args[0].elapsed_s) /* approx scale */,
            n_threads);
     printf("compression: %.1f:1, avg PSNR %.2f dB\n",
-           (double)agg_bytes_in / agg_bytes_enc, psnr_sum / n_done);
+           (double)agg_bytes_in / (double)agg_bytes_enc, psnr_sum / (double)n_done);
 
     for (size_t i = 0; i < n_paths; ++i) free(paths[i]);
     free(paths); free(thr); free(args);
